pyproject.toml
[project]
name = "medicoder"
version = "0.1.0"
description = ""
authors = [
    {name = "Abhiyan Timilsina",email = "abhiyantimilsina@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langgraph (>=1.0.10,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "langsmith (>=0.7.14,<0.8.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-core (>=1.2.17,<2.0.0)",
    "google-cloud-aiplatform (>=1.140.0,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "langgraph-cli[inmem] (>=0.4.14,<0.5.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "langchain-google-vertexai (>=3.2.2,<4.0.0)",
    "polars (>=1.38.1,<2.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "google-cloud-storage (>=3.9.0,<4.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]

[tool.poetry]
packages = [{include = "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.5,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]

[tool.ruff]
# Target version
target-version = "py310"
line-length = 88

[tool.ruff.lint]
# Enable Pyflakes (F), E/W (Pycodestyle), I (isort), B (flake8-bugbear)
select = ["E", "F", "W", "I", "B", "UP", "N", "PLC", "PLE", "PLR", "PLW"]
ignore = [
    "PLR0913", # Allow more than 5 arguments in constructors
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langchain.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that require external services (e.g. LLM endpoints)"
]
filterwarnings = [
    "ignore:Core Pydantic V1 functionality isn't compatible with Python 3.14:UserWarning",
    "ignore:'_UnionGenericAlias' is deprecated:DeprecationWarning",
]
